[#11636] java: provide way to keep cluster data after test
Summary:
Currently, Java tests run clusters in a temporary directory and delete
it afterwards.  It is possible to keep the postgres regress test related
files by specifying TEST_TMPDIR environment variable, but other data
directories like for master and tserver are cleaned up in a separate
part of the Java code.  Address this cleanup with a new environment
variable YB_JAVATEST_KEEPDATA.  When that environment variable is true,
don't do the cleanup of master/tserver data and also some other data
like the process_info directory and core files.

Furthermore, use EnvAndSysPropertyUtil to allow setting system property
yb.javatest.keepdata as an alternative to the environment variable.

Close: #11636

Test Plan:
Jenkins: java only, build platform: alma8, build type: release, compiler: clang12

    #!/usr/bin/env bash
    YB_JAVATEST_KEEPDATA=true ./yb_build.sh --java-test TestPgRegressPgMisc
    # See tmp dir is still there with data
    find /tmp -name 'ybtest-*'
    tree $(find /tmp -name 'ybtest-*')

    # Same as above but using system properties
    YB_EXTRA_MVN_OPTIONS_IN_TESTS='-Dyb.javatest.keepdata=true' ./yb_build.sh --java-test TestPgRegressPgMisc --sj
    find /tmp -name 'ybtest-*'

Reviewers: dmitry, alex

Reviewed By: dmitry, alex

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D15740

+++ java/yb-client/src/test/java/org/yb/minicluster/MiniYBCluster.java
@@ -1006,17 +1006,21 @@ public class MiniYBCluster implements AutoCloseable {
     String processInfoDir = getProcessInfoDir();
     processCoreFiles(processInfoDir);
     pathsToDelete.add(processInfoDir);
-    for (String path : pathsToDelete) {
-      try {
-        File f = new File(path);
-        LOG.info("Deleting path: " + path);
-        if (f.isDirectory()) {
-          FileUtils.deleteDirectory(f);
-        } else {
-          f.delete();
+    if (ConfForTesting.keepData()) {
+      LOG.info("Skipping deletion of data paths");
+    } else {
+      for (String path : pathsToDelete) {
+        try {
+          File f = new File(path);
+          LOG.info("Deleting path: " + path);
+          if (f.isDirectory()) {
+            FileUtils.deleteDirectory(f);
+          } else {
+            f.delete();
+          }
+        } catch (Exception e) {
+          LOG.warn("Could not delete path {}", path, e);
         }
-      } catch (Exception e) {
-        LOG.warn("Could not delete path {}", path, e);
       }
     }
     LOG.info("Mini cluster shutdown finished");
